Command cleanup, github REST API support
